# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Cat F\u00fcd API",
        "version": "0.0.1",
        "description": "The Cat F\u00fcd API: a shared pet-food catalogue (brand ladder, products, barcodes, advisories) plus each user's private pets, pantry, ratings and meal history.\n\nAuthenticate with `POST /register` or `POST /login`, then send `Authorization: Bearer <token>`. Catalogue reads are scoped to the caller's region and\nto what they may see: approved products, plus their own pending contributions. Staff (moderator, admin) see more fields."
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1"
        }
    ],
    "security": [
        {
            "http": []
        }
    ],
    "paths": {
...
        "/admin/products/{id}/moderate": {
            "post": {
                "operationId": "productModeration.moderate",
                "summary": "Approve, send back for changes, or decline a product. Sending back or declining needs a note the contributor can read",
                "tags": [
                    "ProductModeration"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ModerateProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "product": {
                                            "$ref": "#/components/schemas/ProductResource"
                                        }
                                    },
                                    "required": [
                                        "product"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/admin/requests/resolve": {
            "post": {
                "operationId": "requestQueue.resolve",
                "summary": "Place every product in a group on one node. Add `approve` to make them public in the same step; products that cannot\nbe approved (a barcode already in the public catalogue) stay placed and are listed under `failed` for merging",
                "tags": [
                    "RequestQueue"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResolveRequestRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "placed": {
                                            "type": "integer"
                                        },
                                        "approved": {
                                            "type": "integer"
                                        },
                                        "failed": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "reason": {
                                                        "type": "string"
                                                    }
                                                },
                                                "required": [
                                                    "id",
                                                    "reason"
                                                ]
                                            }
                                        }
                                    },
                                    "required": [
                                        "placed",
                                        "approved",
                                        "failed"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
        "/admin/requests/decline": {
            "post": {
                "operationId": "requestQueue.decline",
                "summary": "Decline a whole group, with a reason each contributor can read",
                "tags": [
                    "RequestQueue"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DeclineRequestRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "declined": {
                                            "type": "integer"
                                        }
                                    },
                                    "required": [
                                        "declined"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "http": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "DeclineRequestRequest": {
                "type": "object",
                "properties": {
                    "requested_path": {
                        "type": "string",
                        "maxLength": 500
                    },
                    "note": {
                        "type": "string",
                        "description": "Shown to each contributor.",
                        "maxLength": 2000
                    }
                },
                "required": [
                    "requested_path",
                    "note"
                ],
                "title": "DeclineRequestRequest"
            },
...
            "ModerateProductRequest": {
                "type": "object",
                "properties": {
                    "action": {
                        "type": "string",
                        "enum": [
                            "approve",
                            "needs_changes",
                            "reject"
                        ]
                    },
                    "note": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "Required when sending a product back or declining it; the contributor can read it.",
                        "maxLength": 2000
                    }
                },
                "required": [
                    "action"
                ],
                "title": "ModerateProductRequest"
            },
...
            "ProductBarcodeResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "gtin": {
                        "type": "string"
                    },
                    "upc_a": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "pack_label": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "id",
                    "gtin",
                    "upc_a",
                    "pack_label"
                ],
                "title": "ProductBarcodeResource"
            },
...
            "ProductResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "region": {
                        "type": "string"
                    },
                    "gtin": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "upc_a": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "has_barcode": {
                        "type": "boolean"
                    },
                    "placed": {
                        "type": "boolean"
                    },
                    "brand": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "species": {
                        "type": "string"
                    },
                    "kind": {
                        "type": "string"
                    },
                    "form": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "ingredients": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "nutrition": {
                        "type": [
                            "array",
                            "null"
                        ],
                        "items": {}
                    },
                    "image_url": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "source": {
                        "type": "string"
                    },
                    "last_verified_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "line": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "texture": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "source_url": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "brand_node_id": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "path_text": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "title_as_listed": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "requested_path": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "moderation_status": {
                        "type": "string"
                    },
                    "review_note": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "reviewed_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "formula_version": {
                        "type": "integer"
                    },
                    "formula_changed_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TagResource"
                        }
                    },
                    "barcodes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductBarcodeResource"
                        }
                    },
                    "audit_status": {
                        "type": "string"
                    },
                    "audit_notes": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "import_key": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "meta": {
                        "type": [
                            "array",
                            "null"
                        ],
                        "items": {}
                    },
                    "created_by": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "reviewed_by": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "merged_into_id": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "id",
                    "region",
                    "gtin",
                    "upc_a",
                    "has_barcode",
                    "placed",
                    "brand",
                    "name",
                    "species",
                    "kind",
                    "form",
                    "description",
                    "ingredients",
                    "nutrition",
                    "image_url",
                    "source",
                    "last_verified_at",
                    "line",
                    "texture",
                    "source_url",
                    "brand_node_id",
                    "path_text",
                    "title_as_listed",
                    "requested_path",
                    "moderation_status",
                    "review_note",
                    "reviewed_at",
                    "formula_version",
                    "formula_changed_at",
                    "created_at",
                    "updated_at"
                ],
                "title": "ProductResource"
            },
...
            "ResolveRequestRequest": {
                "type": "object",
                "properties": {
                    "requested_path": {
                        "type": "string",
                        "description": "The ladder text exactly as listed in the queue.",
                        "maxLength": 500
                    },
                    "node_id": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "path": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 500
                    },
                    "approve": {
                        "type": [
                            "boolean",
                            "null"
                        ],
                        "description": "Also approve the products once placed."
                    }
                },
                "required": [
                    "requested_path"
                ],
                "title": "ResolveRequestRequest"
            },
...
            "TagResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "group": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "label": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "group",
                    "slug",
                    "label"
                ],
                "title": "TagResource"
            },
...
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
...
        }
    }
}